+++ packages/shopify-app-remix/src/server/unauthenticated/admin/unauthenticated.admin.doc.ts
@@ -2,11 +2,21 @@ import {ReferenceEntityTemplateSchema} from '@shopify/generate-docs';
 
 const data: ReferenceEntityTemplateSchema = {
   name: 'Unauthenticated admin',
-  description:
-    "Allows interacting with the Admin API on requests that didn't come from Shopify." +
-    '\n\n> Caution: This should only be used for Requests that do not originate from Shopify.' +
-    '\n> You must do your own authentication before using this method.' +
-    "\n>This function doesn't perform **any** validation and shouldn't rely on unvalidated user input.",
+  description: `Allows interacting with the Admin API when working outside of Shopify requests.
+This enables apps to integrate with 3rd party services and perform background tasks.
+
+> Caution:
+> This function doesn't perform **any** validation and shouldn't rely on raw user input.
+
+When using this function, consider the following:
+
+#### Background tasks
+
+Apps should ensure that the shop domain is authenticated when enqueueing jobs.
+
+#### 3rd party service requests
+
+Apps must obtain the shop domain from the 3rd party service in a secure way.`,
   category: 'Unauthenticated',
   type: 'object',
   isVisualComponent: false,
@@ -17,7 +27,7 @@ const data: ReferenceEntityTemplateSchema = {
       type: 'GetUnauthenticatedAdminContext',
     },
   ],
-  jsDocTypeExamples: ['UnauthenticatedAdminContext', 'AdminApiContext'],
+  jsDocTypeExamples: ['UnauthenticatedAdminContext'],
   related: [
     {
       name: 'API context',