+++ packages/shopify-app-session-storage-dynamodb/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @shopify/shopify-app-session-storage-dynamodb
 
+## 3.0.0
+
+### Minor Changes
+
+- 64e0246: Update shopify-api version to 9.2.0
+
+### Patch Changes
+
+- 2b54692: Update @aws-sdk/client-dynamodb from 3.470.0 to 3.504.0.
+- e2a8fd8: Updated the `@aws-sdk/util-dynamodb` dependency to v3.499.0
+- f5742c1: Updated dependency on `@shopify/shopify-api`
+- Updated dependencies [f5742c1]
+- Updated dependencies [64e0246]
+  - @shopify/shopify-app-session-storage@2.1.0
+
 ## 2.0.4
 
 ### Patch Changes